move landing component to page
